requirements.txt
PIL
python-dotenv
requests